requirements.txt
requests~=2.27.1
selenium~=4.1.3
python-telegram-bot~=13.11
webdriver_manager~=3.5.3